internal/gutil: add
