+++ internal/ui/selfpage.go
@@ -10,6 +10,7 @@ import (
 	"strings"
 	"time"
 
+	"deedles.dev/trayscale/internal/gutil"
 	"deedles.dev/trayscale/internal/listmodels"
 	"deedles.dev/trayscale/internal/tsutil"
 	"deedles.dev/trayscale/internal/xnetip"
@@ -73,7 +74,7 @@ type SelfPage struct {
 
 func NewSelfPage(a *App, status *tsutil.IPNStatus) *SelfPage {
 	var page SelfPage
-	fillFromBuilder(&page, selfPageXML)
+	gutil.FillFromUI(&page, selfPageXML)
 	page.init(a, status)
 	return &page
 }
@@ -173,7 +174,7 @@ func (page *SelfPage) init(a *App, status *tsutil.IPNStatus) {
 				dialog.Save(context.TODO(), &a.win.MainWindow.Window, func(res gio.AsyncResulter) {
 					f, err := dialog.SaveFinish(res)
 					if err != nil {
-						if !errHasCode(err, int(gtk.DialogErrorDismissed)) {
+						if !gutil.ErrHasCode(err, int(gtk.DialogErrorDismissed)) {
 							slog.Error("save file", "err", err)
 						}
 						return
